Fix requirements and setup.
The requirements were still defined for Liberty; revise these for
Juno.

The setup.sh in Juno needs to additionally add heat_stack_owner role
to the nios user or that user will not be able to create stacks.

The NS group member resource was using an exception class not
available in Juno when it couldn't find the NS group.